+++ BudgetApp2/BudgetApp2/Expense.cs
@@ -8,7 +8,7 @@ namespace BudgetApp2
     {
         public string CategoryName { get; set; }
         public string CategoryImage { get; set; }
-        public int  CategoryCost { get; set; }
+        public decimal  CategoryCost { get; set; }
         public string CategoryDesc { get; set; }
         public string Date { get; set; }
